Move test helper functions to consolidated codemap testing
